requirements.txt
pyyaml~=6.0.1
diffusers~=0.18.2
thop~=0.1.1-2209072238
torchvision~=0.16.0
numpy~=1.24.3
pillow~=10.2.0
datasets~=2.18.0
requests~=2.31.0
tqdm~=4.66.1
transformers~=4.36.2
huggingface_hub~=0.20.3
packaging~=23.1
bitsandbytes~=0.42.0
pytorch_fid~=0.3.0